requirements.txt
astroid==2.3.3
aws-cfn-bootstrap==2.0
backcall==0.2.0
binance-connector==3.1.1
botocore==1.31.3
certifi==2023.7.22
charset-normalizer==3.2.0
cycler==0.11.0
decorator==5.1.1
Django==2.2.28
docutils==0.14
fonttools==4.38.0
git-remote-codecommit==1.16
idna==3.4
ikp3db==1.4.1
importlib-metadata==6.7.0
ipython==7.34.0
isort==4.3.21
jedi==0.18.2
Jinja2==3.1.2
jmespath==1.0.1
kiwisolver==1.4.4
lazy-object-proxy==1.4.3
lockfile==0.11.0
MarkupSafe==2.1.3
matplotlib==3.5.3
matplotlib-inline==0.1.6
mccabe==0.6.1
numpy==1.21.6
packaging==23.1
pandas==1.3.5
parso==0.8.3
pbr==5.11.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
prompt-toolkit==3.0.39
ptyprocess==0.7.0
pycryptodome==3.18.0
Pygments==2.15.1
pylint==2.4.4
pylint-django==2.3.0
pylint-flask==0.6
pylint-plugin-utils==0.8.2
pyparsing==3.1.0
pystache==0.5.4
python-daemon==2.2.3
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
schedule==1.2.0
simplejson==3.2.0
six==1.16.0
sqlparse==0.4.4
stevedore==3.5.2
traitlets==5.9.0
typed-ast==1.4.3
typing-extensions==4.7.1
urllib3==1.26.16
virtualenv==16.2.0
virtualenv-clone==0.5.7
virtualenvwrapper==4.8.4
wcwidth==0.2.6
websocket-client==1.6.1
wrapt==1.11.2
zipp==3.15.0
